requirements.txt
streamlit>=1.37.0
langgraph>=0.2.35
requests>=2.32.0
pydantic>=2.8.0
tensorflow>=2.16.0
transformers>=4.44.0
torch>=2.3.0
pillow>=10.4.0
numpy>=1.26.0
duckduckgo-search>=6.2.0
python-dotenv>=1.0.0
